# openapi.yaml
openapi: 3.0.3
info:
  title: EasyBooking API
  version: 1.0.0
  description: REST API for the EasyBooking endterm project.
servers:
  - url: http://127.0.0.1:3000
paths:
...
  /api/bookings:
    get:
      summary: List bookings (auth)
      security: [{ cookieAuth: [] }]
      parameters:
        - in: query
          name: scope
          schema:
            type: string
            enum: [mine, all]
        - in: query
          name: status
          schema:
            type: string
            enum: [confirmed, cancelled]
      responses:
        '200':
          description: Bookings list
    post:
      summary: Create booking (auth)
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingCreateInput'
      responses:
        '201':
          description: Created
  /api/bookings/{id}:
    get:
      summary: Get booking by ID (owner/admin)
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Booking details
    put:
      summary: Update booking (owner/admin)
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingUpdateInput'
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete booking (owner/admin)
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Deleted
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: easybook.sid
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
      description: MongoDB ObjectId
...
  schemas:
...
    BookingCreateInput:
      type: object
      required: [hotelId, checkIn, checkOut, guests]
      properties:
        hotelId: { type: string }
        checkIn: { type: string, format: date }
        checkOut: { type: string, format: date }
        guests: { type: integer, minimum: 1, maximum: 10 }
        status: { type: string, enum: [confirmed, cancelled] }
        notes: { type: string }
    BookingUpdateInput:
      type: object
      properties:
        hotelId: { type: string }
        checkIn: { type: string, format: date }
        checkOut: { type: string, format: date }
        guests: { type: integer, minimum: 1, maximum: 10 }
        status: { type: string, enum: [confirmed, cancelled] }
        notes: { type: string }
...